backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
python-multipart
aiofiles

opencv-python
numpy
tensorflow
torch
--extra-index-url https://download.pytorch.org/whl/cu116

git+https://github.com/openai/whisper.git
speechrecognition
pyaudio

anthropic
